# internal/management/openapi/openapi.yaml
components:
...
  schemas:
...
    Activations:
      properties:
        errors:
          items:
            $ref: '#/components/schemas/HostError'
          type: array
        finishedAt:
          description: When collection finished
          format: date-time
          type: string
        items:
          items:
            $ref: '#/components/schemas/Activation'
          type: array
        nextCursor:
          description: Opaque cursor to pass back as `cursor` to fetch the next page,
            omitted on the last page
          type: string
        partial:
          description: True when some hosts could not be queried, so their activations
            are missing
          type: boolean
        startedAt:
          description: When collection started
          format: date-time
          type: string
      required:
      - errors
      - finishedAt
      - items
      - partial
      - startedAt
      type: object
...
    Error:
      description: The body of every error response.
      properties:
        code:
          description: |-
            A machine-readable error code
            See the API description for the meaning of each code
          enum:
          - badRequest
          - unauthorized
          - forbidden
          - notFound
          - notApplicable
          - methodNotAllowed
          - exclusiveLeaseHeld
          - lastServer
          - hostUnavailable
          - hostReattached
          - noHostsReachable
          - eventHistoryDisabled
          - payloadTooLarge
          - timeout
          - internal
          type: string
        details:
          additionalProperties: {}
          description: |-
            Additional machine-readable context, present for some codes:
            `lastServer` has `actorTypes` (array of strings);
            `noHostsReachable` has `errors` (array of `HostError`)
          type: object
        message:
          description: A human-readable description of the error
          type: string
        requestId:
          description: The request ID, also returned in the `X-Request-Id` header
          type: string
        retryable:
          description: Present and `true` when repeating the request later may succeed
          type: boolean
      required:
      - code
      - message
      type: object
...
  securitySchemes:
    bearerAuth:
      description: 'A read-only or management token from the server configuration,
        sent as `Authorization: Bearer <token>`. Read-only tokens grant every scope
        except those ending in `:manage`; management tokens grant every scope.'
      in: header
      name: Authorization
      type: apiKey
info:
  contact: {}
  description: |
    The management API exposes read-only views of a Francis cluster (hosts, runtimes, placements, in-memory activations, actor state, jobs, alarms, and workflows), plus a small set of audited actions (drain a host, deactivate an actor, cancel/suspend/resume a workflow instance).

    ## Authentication and scopes

    Every route under `/api/v1` requires an `Authorization: Bearer <token>` header.
    `GET /api/v1/openapi.yaml` and `GET /healthz` (served at the root, outside `/api/v1`) are public and don't require a token.

    Each operation declares the scope it requires in its description and in the `x-required-scope` extension.
    The server is configured with two token lists:

    - **Read-only tokens** receive every scope except those ending in `:manage`.
    - **Management tokens** receive every scope.

    | Scope | Grants |
    | --- | --- |
    | `cluster:read` | Cluster summary, runtimes, hosts |
    | `actors:read` | Activations, placements, actor types, actor-state listings |
    | `actors:state:read` | Reading the stored state of an actor (audited); a workflow's actor types also need `workflows:data:read` |
    | `workflows:read` | Workflows, instances, event history |
    | `workflows:data:read` | Including workflow instance input and output in instance details, and reading the stored state of a workflow's actor types (audited) |
    | `jobs:read` | Jobs and alarms |
    | `actors:manage` | Deactivating an actor |
...

...

...

...

...

...

...

...

...

...

...
openapi: 3.0.3
paths:
  /api/v1/activations:
    get:
      description: |-
        Requires scope `actors:read`.

        Walks the hosts in host ID order, querying a batch of hosts concurrently, until the page is full.
        Items are ordered by host, then by actor type and actor ID within each host.
        The filters can be repeated, and each host is then queried for one of the actor types after another, in alphabetical order.

        This is a fan-out endpoint: a host that cannot be queried is reported in `errors` and sets `partial: true`, and its activations are missing from the page.
        If every queried host failed, the response is `503 noHostsReachable` with the host errors in `details.errors`.
        Each host is sampled independently, so the listing is not transactionally consistent: an actor that moved between hosts during collection or between pages may be listed twice or not at all.
      operationId: listActivations
      parameters:
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          type: integer
      - description: Opaque cursor returned as nextCursor by the previous page; omit
          for the first page
        in: query
        name: cursor
        schema:
          type: string
      - description: Only return activations of these actor types; repeat the parameter
          for several
        in: query
        name: type
        schema:
          items:
            type: string
          type: array
      - description: Only query these hosts, returning 404 notFound when none of them
          is registered; repeat the parameter for several
        in: query
        name: host
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activations'
          description: A page of activations
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`notFound`: the host in the host parameter is not registered'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`noHostsReachable`: none of the queried hosts could be reached,
            with the host errors in details.errors; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: List in-memory activations across hosts
      tags:
      - Actors
      x-required-scope: actors:read
  /api/v1/actor-states:
    get:
      description: |-
        Requires scope `actors:read`.

        Lists the IDs of the actors of one type that have stored state, ordered by actor ID, with their workflow labels when the state belongs to a workflow orchestrator.
        The state itself is not returned; use `GET /api/v1/actor-states/{type}/{id}`.
      operationId: listActorStates
      parameters:
      - description: The actor type, which must not be empty and must not contain
          a slash
        in: query
        name: type
        required: true
        schema:
          minLength: 1
          type: string
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          type: integer
      - description: Opaque cursor returned as nextCursor by the previous page; omit
          for the first page
        in: query
        name: cursor
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page-ActorStateItem'
          description: A page of actors with stored state
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: List the actors of a type that have stored state
      tags:
      - Actors
      x-required-scope: actors:read
  /api/v1/actor-states/{type}/{id}:
    get:
      description: |-
        Requires scope `actors:state:read`. Every read is audited (the data itself is never logged).
        Reading the state of a workflow's actors, such as an orchestrator's journal, also requires `workflows:data:read`, since it holds the instances' input and output.

        Returns the exact stored MessagePack bytes with `Content-Type: application/msgpack`.
        The stored value is not decoded or converted; error responses use JSON.
      operationId: getActorState
      parameters:
      - description: The actor type, which must not contain a slash
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: The actor ID, which must not contain a slash
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/msgpack:
              schema:
                format: binary
                type: string
          description: The exact stored MessagePack bytes
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          $ref: '#/components/responses/getActorState_400'
        "401":
          $ref: '#/components/responses/getActorState_401'
        "403":
          $ref: '#/components/responses/getActorState_403'
        "404":
          $ref: '#/components/responses/getActorState_404'
        "500":
          $ref: '#/components/responses/getActorState_500'
        "504":
          $ref: '#/components/responses/getActorState_504'
      security:
      - bearerAuth: []
      summary: Read an actor's stored state
      tags:
      - Actors
      x-required-scope: actors:state:read
  /api/v1/actor-types:
    get:
      description: |-
        Requires scope `actors:read`.

        Aggregates the actor types registered by every live host, sorted by actor type, with per-host placement usage and job retention, and the execution capacity groups the hosts report.
        Not paginated.

        This is a fan-out endpoint for the capacity groups: a host that cannot be queried is reported in `errors`, sets `partial: true`, and contributes no `capacityGroups` entries (its registration is still listed under `hosts`).
      operationId: listActorTypes
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorTypes'
          description: The actor types
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: List actor types
      tags:
      - Actors
      x-required-scope: actors:read
  /api/v1/actors/{type}/{id}/deactivate:
    post:
      description: |-
        Requires scope `actors:manage`. The action is audited.

        Looks up the host the actor is actively placed on (without creating a placement), and asks that host to halt the actor, returning `200` once the host acknowledged.
        The actor's stored state is kept; it is re-activated on its next invocation.

        - When the actor is not active anywhere, or its host went away since the lookup, the response is `200` with `notActive: true` and no request is sent to a host.
        - When the host reports the actor was not active, the response is `200` with `notActive: true` and `hostId` set.
        - An exclusive-access lease on the cluster doesn't stop a deactivation, which writes nothing to the provider; the lease makes every host halt its actors anyway.
        - If the host reconnected before it received the request, including to another runtime replica, the request is sent again to its new session; only if the host keeps reconnecting is the response `409 hostReattached`, and the request can be sent again.

        The request body is not read.
      operationId: deactivateActor
      parameters:
      - description: The actor type, which must not contain a slash
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: The actor ID, which must not contain a slash
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivateResponse'
          description: The actor was deactivated, or was not active
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`hostReattached`: the host kept reconnecting, so the request
            could not be delivered to its current session; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`hostUnavailable`: the host, or the runtime owning its session,
            could not be reached or was too busy; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: Deactivate an actor
      tags:
      - Actions
      x-required-scope: actors:manage
...
  /api/v1/token:
    get:
      description: |-
        Requires a valid token, with no particular scope.

        Returns the scopes the bearer token grants, so a client can tell a read-only token from a management token before it offers an action.
      operationId: getToken
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
          description: The scopes of the token
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: Describe the caller's token
      tags:
      - Meta
...